Move penalty related functions from client to penalty scheduler
